added npc segments
