--- Server/MirObjects/NPC_dep/NPCPage.cs
@@ -1,20 +1,0 @@
-﻿//using System;
-//using System.Collections.Generic;
-//using System.Linq;
-//using System.Text;
-//using System.Text.RegularExpressions;
-
-//namespace Server.MirObjects
-//{
-//    public class NPCPage
-//    {
-//        public string Key;
-//        public List<NPCSegment> SegmentList = new List<NPCSegment>();
-//        public List<string> Args = new List<string>();
-
-//        public NPCPage(string key)
-//        {
-//            Key = key;
-//        }
-//    }
-//}
